Refactored scaffolging, updated README
