+++ README.md
@@ -1,67 +1,47 @@
-# Up-time resource monitoring node.js API
+# Up-time resource monitoring node.js app
 This API was built without any npm module!
 It allows a user to enter the URL's they want to be monitored, and receive alerts when this resource go _down_ and _come back_
 
-[Explore API](https://web.postman.co/collections/3368587-2e0ee175-4293-4ee9-9847-f455623a685b?workspace=8358da19-19c9-45ba-8d76-2d6bc9d62598)
+## API
 
+   API is hosting on localhost:***port***/api
 
-#### NOTES:
- - Run every command in root directory
+   [Explore DOCS](https://web.postman.co/collections/3368587-2e0ee175-4293-4ee9-9847-f455623a685b?workspace=8358da19-19c9-45ba-8d76-2d6bc9d62598)
 
- - Scafolding
-    - ```.data``` - here is storing all data about (tokens, checks, users)
+## Run the application
 
-    - ```.logs``` - here is storing logs (name of logs equals to log id) and archived logs
-
-    - ```https``` - put your key and cert to run https server
-
- - Every time you start the app the following will happen
-    - All checks will be immediately executed
-
-    - The loop will be started so the checks will execute later on
-
-    - Each check log file will be compressed immediately
-
-    - Each check log file will be truncated
-
-    - The logs compression loop will be started
-## CLI
-
+#### NOTES: Run every command in root directory
 Start application (default in ***dev*** mode)
 ```javascript
 $ node index.js
+```
+Start application in ***production*** mode
+```console
+$ NODE_ENV=production node index.js
 ```
 
 Start application and view logs from ***server***
 ```console
 $ NODE_DEBUG=server node index.js
-
-Workers have successfully started
-The server is up on port 3000 and running in dev mode
-The server is up on port 3001 and running in dev mode
-SERVER 10795: GET /  200
-SERVER 10795: GET / public/app.js 200
-SERVER 10795: GET / public/app.css 200
-SERVER 10795: GET / public/logo.png 200
 ```
 
 Start application and view logs from ***workers***
 ```console
 $ NODE_DEBUG=workers node index.js
-
-Workers have successfully started
-The server is up on port 3000 and running in dev mode
-The server is up on port 3001 and running in dev mode
-WORKERS 14429: Success truncating log file
-WORKERS 14429: Logging to file succeded
-WORKERS 14429: Outcome has not changed no alert is needed
 ```
 
-Start application in ***production*** mode
-```console
-$ NODE_ENV=production node index.js
+## CLI tools
 
-Workers have successfully started
-The server is up on port 5000 and running in production mode
-The server is up on port 5001 and running in production mode
-```
+Command | Action
+------------ | -------------
+```exit``` | Kill the CLI (and the rest of the application)
+```man``` | Show this help page
+```help``` | Alias of the 'man' command
+```stats``` | Get statistics on the underlying operating system and resource utilization
+```list users``` | Show a list of all the registered (undeleted) users in the system
+```more user info --{userId}``` | Show details of a specified user
+```list checks --up --down``` | "Show a list of all the active checks in the system, including their state. The '--up' and '--down flags are both optional.'",
+```more check info --{checkId}``` | Show details of a specified check
+```list logs``` | Show a list of all the log files available to be read (compressed only)
+```more log info --{logFileName}``` | Show details of a specified log file
+
