+++ lib/cli/index.js
@@ -9,7 +9,7 @@ const debug = util.debuglog("cli");
 const events = require("events");
 
 const Responder = require("./responders");
-const helpers = require("../helpers");
+const helpers = require("../common/helpers");
 
 class _events extends events {}
