+++ lib/cli/responders.js
@@ -1,9 +1,9 @@
 const os = require("os");
 const v8 = require("v8");
 
-const helpers = require("../helpers");
-const _data = require("../data");
-const _logs = require("../logs");
+const helpers = require("../common/helpers");
+const _data = require("../common/data");
+const _logs = require("../common/logs");
 
 const Responders = {
     // Help / Man
