+++ lib/common/data.js
@@ -6,14 +6,14 @@
 const fs = require("fs");
 const path = require("path");
 
-const helpers = require("./helpers");
+const helpers = require("../common/helpers");
 
 // Conctainer for the  module
 
 const lib = {};
 
 // Base directory for the data folder
-lib.baseDir = path.join(__dirname, "/../.data/");
+lib.baseDir = path.join(__dirname, "/../../.data/");
 
 // Write data to a file
 lib.create = function(dir, file, data, callback) {
