+++ lib/common/helpers.js
@@ -10,7 +10,7 @@ const StringDecoder = require("string_decoder").StringDecoder;
 const path = require("path");
 const fs = require("fs");
 
-const config = require("./config");
+const config = require("../config");
 
 const helpers = {};
 
@@ -148,7 +148,7 @@ helpers.getTemplate = function(templateName, data, callback) {
     data = typeof data == "object" && data !== null ? data : {};
 
     if (templateName) {
-        const templatesDir = path.join(__dirname, "/../templates/");
+        const templatesDir = path.join(__dirname, "/../../templates/");
         fs.readFile(templatesDir + templateName + ".html", "utf-8", function(
             err,
             str
@@ -217,7 +217,7 @@ helpers.readStaticAsset = function(fileName, callback) {
         typeof fileName == "string" && fileName.length > 0 ? fileName : false;
 
     if (fileName) {
-        const publicDir = path.join(__dirname, "/../public/");
+        const publicDir = path.join(__dirname, "/../../public/");
         fs.readFile(publicDir + fileName, function(err, data) {
             if (!err && data) {
                 callback(false, data);
