+++ lib/common/logs.js
@@ -11,7 +11,7 @@ const path = require("path");
 // Container for the module
 const lib = {};
 
-lib.baseDir = path.join(__dirname, "/../.logs/");
+lib.baseDir = path.join(__dirname, "/../../.logs/");
 
 // Append a file to a string. Create a file it dosen`t exist
 lib.append = function(file, str, callback) {
