+++ lib/server/controllers/checks.js
@@ -1,6 +1,6 @@
-const _data = require("../data");
-const config = require("../config");
-const helpers = require("../helpers");
+const _data = require("../../common/data");
+const helpers = require("../../common/helpers");
+const config = require("../../config");
 const TokensController = require("./tokens");
 
 const ChecksController = {
