+++ lib/server/controllers/tokens.js
@@ -1,5 +1,5 @@
-const _data = require("../data");
-const helpers = require("../helpers");
+const _data = require("../../common/data");
+const helpers = require("../../common/helpers");
 
 // Tokens controller
 const TokensController = {
