+++ lib/server/controllers/users.js
@@ -1,5 +1,5 @@
-const _data = require("../data");
-const helpers = require("../handles");
+const _data = require("../../common/data");
+const helpers = require("../../common/helpers");
 const TokensController = require("./tokens");
 
 const UsersController = {
