+++ lib/server/handles.js
@@ -1,4 +1,4 @@
-const helpers = require("./helpers");
+const helpers = require("../common/helpers");
 
 // Define all the handlers
 const handlers = {};
