+++ lib/server/index.js
@@ -12,9 +12,9 @@ const path = require("path");
 const StringDecoder = require("string_decoder").StringDecoder;
 const fs = require("fs");
 
-const config = require("./config");
+const config = require("../config");
 const handlers = require("./handles");
-const helpers = require("./helpers");
+const helpers = require("../common/helpers");
 
 const util = require("util");
 const debug = util.debuglog("server");
@@ -29,8 +29,8 @@ server.httpServer = http.createServer(function(req, res) {
 
 // Instantiate https server
 server.httpsServerOptions = {
-    key: fs.readFileSync(path.join(__dirname, "/../https/key.pem")),
-    cert: fs.readFileSync(path.join(__dirname, "/../https/cert.pem"))
+    key: fs.readFileSync(path.join(__dirname, "/../../https/key.pem")),
+    cert: fs.readFileSync(path.join(__dirname, "/../../https/cert.pem"))
 };
 server.httpsServer = https.createServer(server.httpsServerOptions, function(
     req,
