+++ lib/workers/index.js
@@ -11,9 +11,9 @@ const url = require("url");
 const util = require("util");
 const debug = util.debuglog("workers");
 
-const _data = require("./data");
-const _logs = require("./logs");
-const helpers = require("./helpers");
+const _data = require("../common/data");
+const _logs = require("../common/logs");
+const helpers = require("../common/helpers");
 
 //Instantiate workers
 const workers = {};
